refactor explain single series
